+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Added
+* **css:** Layout Viewer ([#1279](https://github.com/dso-toolkit/dso-toolkit/issues/1279))
+
 ### Changed
 * **BREAKING: all packages**. Update dependencies ([#1290](https://github.com/dso-toolkit/dso-toolkit/issues/1290)). Overgestapt van `node-sass` naar `sass` (Dart Sass). **stylesheets zijn niet meer node-sass compatible**
 * **BREAKING: react + react-leaflet:** React dependencies van de hele DSO Toolkit ophogen naar `^17.0.2` ([#1305](https://github.com/dso-toolkit/dso-toolkit/issues/1305)).